+++ trunk/Simulator/2dPlanarScene/NArticulatedPlanarController.cs
@@ -10,15 +10,15 @@ namespace Simples.Simulation.Planar2D
     {
         private static float VELOCITY = 50.0f;
         private NArticulatedPlanar robot;
-        private List<int[]> pointList;
+        private List<double[]> pointList;
         public bool running = false;
 
         public NArticulatedPlanarController(NArticulatedPlanar robot)
         {
             this.robot = robot;
-            pointList = new List<int[]>();
+            pointList = new List<double[]>();
         }
-        public void AddPoint(int[] p)
+        public void AddPoint(double[] p)
         {
             pointList.Add(p);
         }
@@ -29,20 +29,20 @@ namespace Simples.Simulation.Planar2D
 
         public void Update(GameTime gameTime)
         {
-            float maxIncrement = (float)gameTime.ElapsedGameTime.TotalSeconds * VELOCITY;
+            double maxIncrement = gameTime.ElapsedGameTime.TotalSeconds * VELOCITY;
             if (running)
             {
                 if (pointList.Count > 0)
                 {
-                    float delta1 = pointList[0][0] - robot.Mechanism.Joints[0].Value;
-                    float delta2 = pointList[0][1] - robot.Mechanism.Joints[1].Value;
-                    float delta3 = pointList[0][2] - robot.Mechanism.Joints[2].Value;
-                    float next1;
-                    float next2;
-                    float next3;
+                    double delta1 = pointList[0][0] - robot.Mechanism.Joints[0].Value;
+                    double delta2 = pointList[0][1] - robot.Mechanism.Joints[1].Value;
+                    double delta3 = pointList[0][2] - robot.Mechanism.Joints[2].Value;
+                    double next1;
+                    double next2;
+                    double next3;
 
-                    float factor;
-                    float max = Math.Max(Math.Abs(delta1), Math.Abs(delta2));
+                    double factor;
+                    double max = Math.Max(Math.Abs(delta1), Math.Abs(delta2));
                     max = Math.Max(max, Math.Abs(delta3));
                     if (max == Math.Abs(delta1))
                     {
@@ -66,7 +66,7 @@ namespace Simples.Simulation.Planar2D
                         next2 = delta2 * factor;
                     }
 
-                    if ((next1 == float.NaN) || (next2 == float.NaN) || (next3 == float.NaN))
+                    if ((next1 == double.NaN) || (next2 == double.NaN) || (next3 == double.NaN))
                     {
                         next1 = 0;
                         next2 = 0;
@@ -77,9 +77,9 @@ namespace Simples.Simulation.Planar2D
                     robot.Mechanism.Joints[1].Value += next2;
                     robot.Mechanism.Joints[2].Value += next3;
 
-                    if ((Math.Abs(robot.Mechanism.Joints[0].Value - pointList[0][0]) < 0.1f) &&
-                        (Math.Abs(robot.Mechanism.Joints[1].Value - pointList[0][1]) < 0.1f) &&
-                        (Math.Abs(robot.Mechanism.Joints[2].Value - pointList[0][2]) < 0.1f))
+                    if ((Math.Abs(robot.Mechanism.Joints[0].Value - pointList[0][0]) < 0.1) &&
+                        (Math.Abs(robot.Mechanism.Joints[1].Value - pointList[0][1]) < 0.1) &&
+                        (Math.Abs(robot.Mechanism.Joints[2].Value - pointList[0][2]) < 0.1))
                     {
                         robot.Mechanism.Joints[2].Value = pointList[0][2];
                         robot.Mechanism.Joints[1].Value = pointList[0][1];
